% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosshairs.R
\name{crosshairs}
\alias{cross-hairs}
\alias{crosshairs}
\title{Creates a Cross-hairs Plot.}
\usage{
crosshairs(x, y, xse, yse, x_lab = NULL, y_lab = NULL, main_lab = NULL,
  confint = 0.95, mod = NULL, mod_lab = NULL, mod_lab_pos = NULL,
  lab_size = 14, pnt_size = 3, file_name = NULL, whis_on = TRUE,
  annotate = NULL, annotate_pos = NULL, grid_dense = FALSE,
  bxplts = TRUE, file_dim = NULL)
}
\arguments{
\item{x}{Effect sizes that will be used in x axis}

\item{y}{Effect sizes that will be used in y axis}

\item{xse}{Standard errors of effect sizes that is used in x axis}

\item{yse}{Standard errors of the effect sizes that is used
in y axis}

\item{x_lab}{Title of the x-axis}

\item{y_lab}{Title of the y-axis}

\item{main_lab}{Main label of the cross-hairs plot.}

\item{confint}{Confidence interval that is used to determine
length of the whiskers.}

\item{mod}{Whether there is a moderator variable?}

\item{mod_lab}{Label of the moderator variable.}

\item{mod_lab_pos}{Determine the positon of the moderator labels.}

\item{lab_size}{Size of the axis titles.}

\item{pnt_size}{Determines the size of point in the plot.}

\item{file_name}{Default name of the resulting bitmap file (e.g.,
CrosshairsOutputFile.png, CrosshairsOuputFile.tiff, etc)}

\item{whis_on}{Whiskers on or off?}

\item{annotate}{If true, mean effect size and correlation between
effect sizes will be printed within the graph.}

\item{annotate_pos}{A value that determine horizontal position
of the annotation.}

\item{grid_dense}{When changed to FALSE, a default denser gridlines
will be used.}

\item{bxplts}{Swithes boxplots on or off.}

\item{file_dim}{An integer vector of length two which is indicating
dimensions of resulting tiff file.}
}
\value{
Creates a \code{.tiff} file in the current working
       directory that you can find with \code{\link[base]{getwd}}.
}
\description{
When given effect sizes and standard errors, creates cross-hairs plots.
}
\details{
\code{crosshairs}'s basic output is a bitmat file that contains cross-hairs
plot of given dependent effect sizes.
}
\examples{
\dontrun{
# Load and attach metaplotr package to the Global Environment
library(metaplotr)

# Remove all variables in the .GlobalEnv, effectively clearing .GlobalEvn
rm(list = ls())

# Find out more about the data set from Ferguson and Brannick (2012)
help("FergusonBrannick2012")

attach data frame to working environment.
attach(FergusonBrannick2012)

# You can check out help file of the \\code{crosshairs} function.
help(crosshairs)

# Basic usage of the \\code{crosshairs} function. You should see a \\code{tiff}
file in your current workind directory after running this command.
crosshairs(pub_z, dis_z, pub_z_se, dis_z_se)

# whis_on option opens and closes whiskers.
crosshairs(pub_z, dis_z, pub_z_se, dis_z_se, whis_on = FALSE)

# confint option can control whiskers length.
crosshairs(pub_z, dis_z, pub_z_se, dis_z_se, confint = .95)
crosshairs(pub_z, dis_z, pub_z_se, dis_z_se, confint = .7)
crosshairs(pub_z, dis_z, pub_z_se, dis_z_se, confint = .3)

# Name of the resulting \\code{tiff} file can be changed.
crosshairs(pub_z, dis_z, pub_z_se, dis_z_se,
file_name = 'A_Different_Name')

# Main and axes labels can be changed.
crosshairs(pub_z, dis_z, pub_z_se, dis_z_se,
file_name = 'A_Different_Main_Label',
main_lab = 'Different Main Label', x_lab = 'Different X',
y_lab = 'Different Y')

# Annotated correlation and meand values can be added to the graph.
crosshairs(pub_z, dis_z, pub_z_se, dis_z_se,
file_name = 'Annotated_Graph', annotate = TRUE,
main_lab = 'Annotated Graph')

# Boxplots can be hidden.
crosshairs(pub_z, dis_z, pub_z_se, dis_z_se,
file_name = 'Hidden_Boxplots', main_lab = 'No Boxplots',
bxplts = FALSE)

# Dimensions of the resulting bitmap file can be changed.
crosshairs(pub_z, dis_z, pub_z_se, dis_z_se,
file_name = 'Smaller_Bitmap_Dimensions', main_lab = 'Smaller Plot',
bxplts = FALSE, file_dim = c(600, 600))
}

}

