% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FergusonBrannick2012}
\alias{FergusonBrannick2012}
\title{Effect sizes and standard errors from Ferguson and Brannick (2012).}
\format{A data frame with 24 rows and 5 variables:
\describe{
  \item{pub_z}{Publication effect size in z scores}
  \item{dis_z}{Dissertation effect size in z scores}
  \item{pub_z_se}{Standard error of the publication effect size z scores}
  \item{dis_z_se}{Standard error of the dissertation effect size z scores}
  \item{mod}{Moderator variable}
}}
\source{
\url{http://dx.doi.org/10.1037/a0027128}
}
\usage{
FergusonBrannick2012
}
\description{
This data set contains Z scores and standards errors from Ferguson and
and Brannick (2012) meta-analysis.
}
\keyword{datasets}

