% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenderDiff01}
\alias{GenderDiff01}
\title{Gender Differences Data (1).}
\format{A data frame with 45 rows and 14 variables:
\describe{
  \item{author}{Author of the publication.}
  \item{pub_year}{Year of the publication.}
  \item{boys_n}{Sample size of male group.}
  \item{boys_r}{Correlation of male group.}
  \item{boys_z}{Effect size of male group.}
  \item{boys_se}{Standard error of the effect size.}
  \item{girls_n}{Sample size of female group.}
  \item{girls_r}{Correlation of female group.}
  \item{girls_z}{Effect size of female group.}
  \item{girls_se}{Standard error of the effect size.}
}}
\source{
\url{http://dx.doi.org/10.1037/a0027128}
}
\usage{
GenderDiff01
}
\description{
This is a data set that show gender differences in effect sizes.
}
\keyword{datasets}

