% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenderDiff02}
\alias{GenderDiff02}
\title{Gender Differences Data (2).}
\format{A data frame with 27 rows and 13 variables:
\describe{
  \item{author}{Author(s) of the study.}
  \item{year}{Publication year.}
  \item{men_n}{Number of male participants.}
  \item{men_r}{Correlation.}
  \item{men_z}{Z-scores.}
  \item{men_se}{Standard error of male participants.}
  \item{women_n}{Number of female participants.}
  \item{women_r}{Correlation.}
  \item{women_z}{Standard effect size.}
  \item{women_se}{Standard error of female participants.}
  \item{menage_m}{To be written.}
  \item{country}{Country from which study was conducted.}
  \item{region}{Continent of the study.}
}}
\source{
\url{http://dx.doi.org/10.1037/a0027128}
}
\usage{
GenderDiff02
}
\description{
This data set contains Z scores and standards errors from Ferguson and
and Brannick (2012) meta-analysis.
}
\keyword{datasets}

