% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenderDiff03}
\alias{GenderDiff03}
\title{Gender Differences Data (3).}
\format{A data frame with 24 rows and 5 variables:
\describe{
  \item{author}{Author of the study.}
  \item{year}{Publication year.}
  \item{men_n}{Number of men.}
  \item{men_r}{Correlation among men.}
  \item{men_z}{Z-scores of men.}
  \item{men_se}{Standard errors of men.}
  \item{gender}{Factor indicating gender.}
}}
\source{
\url{http://dx.doi.org/10.1037/a0027128}
}
\usage{
GenderDiff03
}
\description{
To be written.
}
\keyword{datasets}

