% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{McLeod2007}
\alias{McLeod2007}
\title{Effect sizes from McLeod, Weisz, and Wood (2007).}
\format{A data frame with 45 rows and 14 variables:
\describe{
  \item{author}{Author(s) of the publication used in the meta-analysis}
  \item{year}{Year of the publication}
  \item{n}{Sample size of the primary study used in the meta-analysis}
  \item{age}{Mean age of the participants}
  \item{parent}{Which parent was used? Mother (M), father (F), both (B)}
  \item{techp}{Technology of the parenting measure: questionnaire (Q), interview (I), observation (O)}
  \item{informp}{Who is the informant? Child (C), parent (P), observer (O)}
  \item{dx}{Diagnostic status: yes (Y), no (N)}
  \item{tf}{Timeframe: current symptomatology (C), lifetime symptomatology (L), unspecified (U)}
  \item{techd}{To be added}
  \item{informd}{Information technology used}
  \item{r}{Mean r}
  \item{z}{Effect size as Z-score.}
  \item{var}{Variance of the Z-score.}
}}
\source{
\url{http://dx.doi.org/10.1016/j.cpr.2007.03.001}
}
\usage{
McLeod2007
}
\description{
This data set contains various factors and variables from McLeod, Weisz, and
Wood (2007) Examining the association between parenting and childhood depression: A meta-analysis.
}
\keyword{datasets}

