% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplotr.R
\docType{package}
\name{metaplotr-package}
\alias{metaplotr}
\alias{metaplotr-package}
\title{Creates Cross-hairs Plot.}
\description{
In its current state this package only creates a cross-hairs plots that
makes it possible to compare dependent effects sizes in meta-analysis. This
package was created as a supporting code of the paper Brannick and Gultas
(2016). In the future it will contain functions to create regularly used
plots in meta-analyses studies (i.e., forest, funnel, le abbe, radial, baujat,
GOSH, L'Abbe)
}

