% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sweeney2015}
\alias{Sweeney2015}
\title{Effect sizes from Sweeney and Moyer (2015).}
\format{A data frame with 10 rows and 6 variables:
\describe{
  \item{author}{Author of the primary study}
  \item{year}{Publication of the primary study}
  \item{inten_d}{Effect size of intentions}
  \item{inten_se}{Standard error of intentions effect size}
  \item{beh_d}{Effect size of behaviors}
  \item{beh_se}{Standard error of behaviors effect size}
}}
\source{
\url{http://dx.doi.org/10.1037/hea0000110}
}
\usage{
Sweeney2015
}
\description{
Self-affirmation and responses to health messages: A meta-analysis on
intentions and behavior.
}
\keyword{datasets}

